trivectorai/backend/requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Validation
pydantic==2.7.0
python-dotenv==1.0.1

# AI / LLM
google-generativeai==0.5.4

# Market data
yfinance==0.2.40
pandas==2.2.2

# Backtest engine
vectorbt==0.26.1
numpy==1.26.4

# Utilities
httpx==0.27.0
